requirements.txt
Flask
Werkzeug
vertexai
google-cloud-storage
google-auth
